Delete duplicate file accidentally pushed
